refactor type information and improve design further
