+++ client/src/components/PendingFriendRequests.tsx
@@ -70,7 +70,6 @@ export const PendingFriendRequests = () => {
 
   return (
     <div>
-      <h2 className="text-2xl font-bold mb-4">Pending Friend Requests</h2>
       {loading ? (
         <p>Loading...</p>
       ) : error ? (
