+++ client/src/pages/FriendsPage.tsx
@@ -1,14 +1,22 @@
 import { UserSearch } from "src/components/UserSearch";
-import { FriendList } from "src/components/FriendsharingList";
+import { FriendsPrivacySettings } from "src/components/FriendsharingList";
 import { PendingFriendRequests } from "src/components/PendingFriendRequests";
+import { MapWithFriendMarkers } from "src/components/MapWithFriendMarkers";
 
 export const FriendsPage = () => {
   return (
-    <div className="pt-24 px-4 sm:px-6 lg:px-8 ">
+    <div className="py-24 px-4 sm:px-6 lg:px-8 ">
       <h1>Friends page</h1>
 
-      <FriendList />
+      <MapWithFriendMarkers />
+
+      <h2 className="text-2xl font-bold mb-4">Location Privacy Settings</h2>
+      <FriendsPrivacySettings />
+
+      <h2 className="text-2xl font-bold mb-4">Add a New Friend</h2>
       <UserSearch />
+
+      <h2 className="text-2xl font-bold mb-4">Pending Friend Requests</h2>
       <PendingFriendRequests />
     </div>
   );
